# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Claw Admin API
  description: Unified admin control plane for LLM/Skill/MCP config, SSO, organizations, users, and logs.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path
...
security:
  - BearerAuth: []
paths:
...
  /admin/users/{id}/roles:
    get:
      tags: [admin-users]
      summary: Get user roles per organization
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of { orgId, orgSlug, role }
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [admin-users]
      summary: Set user role in an organization
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [orgId, role]
              properties:
                orgId: { type: string }
                role: { type: string, enum: [owner, admin, member, viewer] }
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'

  # ---------- Admin SSO ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      description: Tenant slug (optional; can come from JWT org_id)
      schema:
        type: string
        example: default
...
  responses:
...
    Forbidden:
      description: Forbidden (admin role required)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: FORBIDDEN }
                  message: { type: string }
...